qca-edma : Change 'echo' to 'echo -n' in qca-edma script
The builtin echo command from the default shell in QRDK- bash
(unlike busybox in QSDK) on musl systems, throws errors in
write operations into procfs and sysfs entries as below due
to a newline character being appended in the end and handled
differently in musl.

'echo: write error: Invalid argument' due to which the
qca-edma service was failing.

Replacing 'echo' with 'echo -n' helps to overcome the error

Change-Id: Ica10a10c1c210d3123b37f86d7e3662dea8a71c1
Signed-off-by: Sandhya <sandhyks@codeaurora.org>